requirements.txt
alembic==1.18.4
bcrypt==4.3.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.2.1
cryptography==46.0.6
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-auth==2.49.1
greenlet==3.2.4
gunicorn==25.3.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.1.post0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.2
pytz==2025.2
razorpay==2.0.1
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.6.3
utils==1.0.2
Werkzeug==3.1.3
WTForms==3.2.1
